Refactor separate switch component into nested switch component
